[Model-Bringup] Consolidate DiffusionGemma tests
Move reusable coverage into existing generic suites and remove redundant configuration-only test files while preserving model-specific regression checks.
