+++ tests/test_module/test_matrix_expansion.py
@@ -7,8 +7,27 @@
 from __future__ import annotations
 
 from test_module.test_categorization_system.suite_loader import (
+    load_server_tests_config,
     load_suite_files_by_category,
 )
+
+
+def test_diffusiongemma_llm_suite_uses_model_specific_conformance_test():
+    suites = load_suite_files_by_category("llm")
+    suite = next(
+        suite for suite in suites if suite["id"] == "diffusiongemma-26b-a4b-it-p300x2"
+    )
+
+    assert suite["weights"] == ["diffusiongemma-26B-A4B-it"]
+    assert suite["device"] == "p300x2"
+    assert [case["template"] for case in suite["test_cases"]] == [
+        "VLLMDiffusionGemmaParamConformanceTest"
+    ]
+
+    templates = load_server_tests_config()["test_templates"]
+    template = templates["VLLMDiffusionGemmaParamConformanceTest"]
+    assert template["module"] == "test_module.llm_tests.vllm_param_conformance_test"
+    assert {"param", "e2e", "slow", "heavy"} <= set(template["markers"])
 
 
 class TestImageMatrixExpansionSDXL:
